requirements.txt
asgiref==3.11.0
asttokens==3.0.1
celery==5.3.4
colorama==0.4.6
decorator==5.2.1
Django==6.0.1
django-extensions==4.1
django-filter==25.2
django-modeltranslation==0.19.13
djangorestframework==3.15.2
markdown==3.7
PyYAML==6.0.3
inflection==0.5.1
uritemplate==4.2.0
executing==2.2.1
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.5
pillow==12.1.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
psycopg2-binary==2.9.10
Pygments==2.19.2
redis==5.0.1
SHop==0.0.2
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.3
wcwidth==0.2.14

